+++ docs/source/whats_new.rst
@@ -51,6 +51,7 @@ Bugs
 - MAINT updating the packages pre-release (:gh:`578` by `Bruno Aristimunha`_)
 - Updating the parameters of the SSVEP_TRCA method (:gh:`589` by `Bruno Aristimunha`_)
 - Fix and updating the parameters for the benchmark function (:gh:`588` by `Bruno Aristimunha`_)
+- Fix :class:`moabb.datasets.preprocessing.SetRawAnnotations` setting incorrect annotations when the dataset's interval does not start at 0 (:gh:`607` by `Pierre Guetschel`_)
 
 
 API changes